remark-abbr: deprecating rehype-abbr
