--- packages/rehype-abbr/.babelrc
@@ -1,9 +1,0 @@
-{
-  "presets": ["es2015"],
-  "plugins": [],
-  "env": {
-    "development": {
-      "sourceMaps": "inline"
-    }
-  }
-}
