--- packages/rehype-abbr/.eslintignore
@@ -1,3 +1,0 @@
-coverage
-public/*
-packages/*/dist/*
