--- packages/rehype-abbr/.npmignore
@@ -1,7 +1,0 @@
-/index.js
-/__tests__
-/.npmignore
-/.nyc_output
-/coverage
-*.log
-/src
