+++ packages/remark-abbr/__tests__/index.js
@@ -3,7 +3,6 @@ import unified from 'unified'
 import reParse from 'remark-parse'
 import stringify from 'rehype-stringify'
 import remark2rehype from 'remark-rehype'
-import rehypeAbbr from 'rehype-abbr'
 
 import remarkAbbr from '../src/'
 
@@ -12,7 +11,6 @@ const render = text => unified()
   .use(remarkAbbr)
   .use(remark2rehype)
   .use(stringify)
-  .use(rehypeAbbr)
   .processSync(text)
 
 it('renders references', () => {
